refactor of routes
