+++ go.mod
@@ -2,4 +2,4 @@ module github.com/violenti/server-pipeline
 
 go 1.16
 
-require github.com/gorilla/mux v1.8.0 // indirect
+require github.com/gorilla/mux v1.8.0
