+++ main.go
@@ -2,22 +2,19 @@ package main
 
 import (
 	"fmt"
-	"io/ioutil"
-	"log"
 	"net/http"
-	"os"
 	"time"
-
+  "github.com/violenti/server-pipeline/route"
 	"github.com/gorilla/mux"
 )
 
-var GITLAB_TOKEN = os.Getenv("GITLAB_TOKEN")
-var URI = os.Getenv("GITLAB")
+//var GITLAB_TOKEN = os.Getenv("GITLAB_TOKEN")
+//var URI = os.Getenv("GITLAB")
 
 func main() {
 	r := mux.NewRouter()
-	r.HandleFunc("/master/{file}", Master).Methods("GET")
-	r.HandleFunc("/development/{file}", Development).Methods("GET")
+	r.HandleFunc("/master/{file}", route.Master).Methods("GET")
+	r.HandleFunc("/development/{file}", route.Development).Methods("GET")
 	http.Handle("/", r)
 
 	srv := &http.Server{
@@ -30,52 +27,4 @@ func main() {
 	}
 	fmt.Println("server-pipeline running on http://0.0.0.0:8000 ")
 	srv.ListenAndServe()
-}
-
-func Master(w http.ResponseWriter, r *http.Request) {
-
-	var varrequest = mux.Vars(r)
-	var FILE = varrequest["file"]
-	var FULL_URI = URI + FILE + "/raw?ref=master" + "&private_token=" + GITLAB_TOKEN
-
-	resp, err := http.Get(FULL_URI)
-	if err != nil {
-		log.Fatalln(err)
-		fmt.Printf("%s", err)
-		os.Exit(1)
-	} else {
-		defer resp.Body.Close()
-		contents, err := ioutil.ReadAll(resp.Body)
-		if err != nil {
-			log.Fatalln(err)
-			os.Exit(1)
-		}
-		fmt.Fprint(w, string(contents))
-		fmt.Print(FULL_URI)
-
-	}
-}
-
-func Development(w http.ResponseWriter, r *http.Request) {
-
-	var varrequest = mux.Vars(r)
-	var FILE = varrequest["file"]
-	var FULL_URI = URI + FILE + "/raw?ref=develop" + "&private_token=" + GITLAB_TOKEN
-
-	resp, err := http.Get(FULL_URI)
-	if err != nil {
-		log.Fatalln(err)
-		fmt.Printf("%s", err)
-		os.Exit(1)
-	} else {
-		defer resp.Body.Close()
-		contents, err := ioutil.ReadAll(resp.Body)
-		if err != nil {
-			log.Fatalln(err)
-			os.Exit(1)
-		}
-		fmt.Fprint(w, string(contents))
-		fmt.Print(FULL_URI)
-
-	}
 }
